fix theme and mode import
