--- packageBAK.json
@@ -1,14 +1,0 @@
-{
-	"name": "Yarn",
-	"version": "0.2.1",
-	"main": "app/index.html",
-	"window": {
-		"toolbar": false,
-		"position": "center",
-		"width": 1280,
-		"height": 720,
-		"min_width": 800,
-		"min_height": 600,
-		"icon": "Yarn.png"
-	}
-}
