+++ src/js/index.js
@@ -5,18 +5,18 @@ window.ko = ko;
 window.$ = window.jQuery = require("jquery");
 require("jquery-contextmenu");
 require("jquery-mousewheel");
-import "./libs/knockout.ace.js";
 
-window.ace = require("ace-builds/src-noconflict/ace");
+import ace from "ace-builds/src-noconflict/ace"
+window.ace = ace//require("ace-builds/src-noconflict/ace");
 ace.config.set("basePath", "libs"); //needed to import yarn mode
 window.define = ace.define;
+
+// import 'ace-builds/webpack-resolver';
+
+
 require("ace-builds/src-min-noconflict/ext-language_tools");
 require("ace-builds/src-min-noconflict/ext-searchbox");
-require("./libs/theme-yarn.js");
-require("./libs/mode-yarn.js");
-// ace.require("./libs/theme-yarn.js");
-// ace.require("./libs/mode-yarn.js");
-
+require("./libs/knockout.ace.js");
 require("jquery.transit");
 
 require("spectrum-colorpicker");
